Examples: Move EditorControls to editor/js.
