+++ editor/index.html
@@ -16,7 +16,6 @@
 
 		<script src="../build/three.js"></script>
 
-		<script src="../examples/js/controls/EditorControls.js"></script>
 		<script src="../examples/js/controls/TransformControls.js"></script>
 
 		<script src="../examples/js/libs/chevrotain.min.js"></script> <!-- VRML -->
@@ -94,6 +93,7 @@
 
 		<script src="../examples/js/vr/WebVR.js"></script>
 
+		<script src="js/EditorControls.js"></script>
 		<script src="js/Storage.js"></script>
 
 		<script src="js/Editor.js"></script>
